Examples
remove token, parser, manipulator out of folder
