+++ helper/json_files/example2.json
@@ -1,0 +1,18 @@
+{
+  "employees": [
+    {
+      "name": "John Doe",
+      "age": 30,
+      "email": "johndoe@example.com",
+      "message": "simpler non-flash version\\",
+      "body": "I think find that this feature is not commonly used.\r\n\r\n## more correct notion of \"sanitization\"\r\n\r\nCurrently sanitization olink and I'll start submitting patches.\r\n\r\n\r\n/cc @xuanxu @tenderlove @tardate ",
+      "prizes": {}
+    },
+    {
+      "name": "Jane Doe",
+      "age": 25,
+      "email": "janedoe@example.com",
+      "married": false
+    }
+  ]
+}
