+++ include/cjson.h
@@ -1,3 +1,8 @@
+#include "json_token.h"
+#include "json_parser.h"
+#include "json_manipulator.h"
+#include "utils.h"
+
 json_element_t *parse_json(char *input);
 void *clean_json_data(json_element_t *node);
-char *get_json_data_in_string(json_element_t *json_node, char *string);
+char *json_to_string(json_element_t *json_node, char *string);
