+++ include/utils.h
@@ -1,3 +1,4 @@
 char *read_file(char *filename);
 int get_percentage_value(int value, int percentage);
 char *num_to_str(int num);
+void write_to_file(char *buffer, char *file_name);
