--- test/tokenization/Makefile
@@ -1,5 +1,0 @@
-example: test_example1.c
-	gcc -o hello test_example1.c
-
-clean:
-	rm hello
