--- test/tokenization/test_example1.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-// #include "json_token.h"
-
-int main() {
-  printf("hii");
-  return 0;
-}
